Fix memory leak in icehelp.cc
Originally mentioned here: https://bugzilla.novell.com/show_bug.cgi?id=546806
Memory leaks should be fixed even if it's right before the program exits.
It's not professional and it's bad practice to rely on the operating system
as not all will free it.  Maybe I'm missing something, but I think it's
'better to be safe than sorry' is the way I see it.
Also see here: http://stackoverflow.com/questions/273209/are-memory-leaks-ever-ok