Fix format buffer overflow (Debian bug 765965) and other issues
Instead of multiple small buffers and string joining on heap, now use
one format buffer and track the string lengts while filling it.

Also fixes reading of cpu frequence which is not available of the sysfs
file used in icewm-git fork, now it falls back to the original location
if needed.