Remove the broken locale switching when retrieving fallback font
As suggested by the code comments. It was broken all the time since
setlocale is not reentrant (the string buffer is rewritten, original
setting gets lost). The point of the code is also not clear; it might
save you a second of load time on a PC from the nighties but what for?
If you set the locale, you want a font that supports this charset!