Stop misusing dynamic allocator as enumeration manager
The YAction class was a hack to manage identities of predefined and
dynamically generated actions. However, this came with runtime
allocation costs where simple enumberation is sufficient for static IDs
or a global counter for dynamic ids.

The type is still pointer-sized and not int for the few cases where the
pointer of derived class was used as key.