metering to debug #184
From the looks of it a lot of pixmaps that are generated with the
root window depth are created and Icewm just expects frames to have
the same depth as the root window.

What will have to be done is to create a YFrameDecor window as a
child of YFrameWindow with a default visual and depth so that frame
decorations can be moved to that window.  It only needs to be placed
below all of the other children and things will display correctly.

This is non-trivial and requires creating a new class that takes over
all the graphics work of the actual frame.