Allow switching and reordering taskbar buttons with keyboard.
Since commit 556b5356 windows can be switched in taskbar order using the mouse wheel.
They taskbar buttons can also be rearranged with drag and drop.

This patch adds the possibility to perform these actions using the keyboard.

This is a forward port of this old patch of mine:
https://sourceforge.net/p/icewm/patches/193/

The new keys are set as this by default:

	KeyTaskBarSwitchNext="Alt+1"
	KeyTaskBarSwitchPrev="Alt+grave"
	KeyTaskBarMoveNext="Alt+Shift+1"
	KeyTaskBarMovePrev="Alt+Shift+grave"

I have been using this setup for over a decade now and find it very convenient
that "grave" and "1" are next to each other and right above the tab key, so that both
swithing methods Alt+Tab and Alt+1/grave are next to each other.