Convert string option passing to mstring_view
Passing that as mstring by value did not make sense since it's split
into tokens down the line anyway. Corefont does not benefit from it but
it's not worse than before (hard trade-off).

Also fix a crash when mixing core and xft fonts, by dropping the local
array handling and replacing with YArray.