Fix building on a case-insensitive file system (#601)
When compiling on a case-insensitive file system, adding the root
directory to the means that C++ headers including the stanard library
header <version> end up including the VERSION file in the root directory
instead. This include path does not appear to be not needed so we can
just drop it.