Section name bugfix and --activate implementation, etc.
Fix the storage of the section name as the explicit "match-name" value.

Drop most remains of the the experimental callback code

Repurpose --apply code as --activate, thus allowing to run the xrandr actiation
command directly from the command line.

Explaining the creation of the label from EDID.

Option to access auto-determined label in the INI file.

Add the %PRIMFLAG even to the auto-determined label strings.