When focusing and RaiseOnFocus, release button grab for issue #759.
Related to issue #742: Strange Firefox Behavior in IceWM. Icewm retains a button grab after focusing. This is problematic for Firefox and Chrome with hover elements on webpages. Data used by canRaise() and overlapped() functions is only updated after YFrameWindow::setWinFocus() ends. When RaiseOnFocus is true we know that the focused window will eventually be raised and it is unnecessary to retain the button grab. This issue remains for RaiseOnClickClient when RaiseOnFocus is false.