Avoid autoreconf failure on existence of nasty tags
If there was an additional tag for the same commit which was preferred
by git (by ordering) and this tag had a namespace (folder) in it, this
condition made autoreconf internal checks with gnits fail.

Also, the errors were not correctly reported due to not running the
shell in e-mode.